Casting warnings

+++ source/room/RoomManager.cpp
@@ -252,8 +252,8 @@ Room* RoomManager::CreateRandomRoom(Room* pRoomConnection, eDirection connectedD
 	int numRoomTries = 0;
 	while(overlapsExistingRoom == true && numRoomTries < 1)
 	{
-		roomLength = (int)(GetRandomNumber(50, 140, 2) * 0.1f);
-		roomWidth = (int)(GetRandomNumber(50, 140, 2) * 0.1f);
+		roomLength = (float)(int)(GetRandomNumber(50, 140, 2) * 0.1f);
+		roomWidth = (float)(int)(GetRandomNumber(50, 140, 2) * 0.1f);
 		roomHeight = 1.0f;
 
 		*randomLengthOffset = GetRandomNumber(-100, 100, 2) * 0.01f;
